Test controller and that work!

+++ src/main/java/repos/UserForumRepository.java
@@ -4,6 +4,16 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserForumRepository extends CrudRepository<UserForum, Long>{
+public interface UserForumRepository extends CrudRepository<UserForum, Integer> {
 
 }
+
+//@NoRepositoryBean
+//public interface DAORepository<T, ID extends Serializable> extends CrudRepository<T, ID>{
+//
+//}
+//
+//@Repository
+//interface UserForumDAOImpl extends DAORepository<UserForum, Long>{	
+//
+//}
